move pending winners type to lib
